+++ src/framework/event.h
@@ -10,7 +10,7 @@
 namespace tram {
 
 typedef uint16_t event_t;
-typedef uint64_t id_t;
+typedef uint32_t id_t;
 typedef uint64_t listener_t;
 
 class EntityComponent;
